Split Android local repositories by domain
