+++ apps/android/app/src/main/java/com/flashcardsopensourceapp/app/di/AppGraph.kt
@@ -52,7 +52,7 @@ import com.flashcardsopensourceapp.data.local.review.SharedPreferencesReviewPref
 import com.flashcardsopensourceapp.data.local.review.SharedPreferencesStoreReviewRequestStore
 import com.flashcardsopensourceapp.data.local.review.StoreReviewRequestStore
 import com.flashcardsopensourceapp.data.local.repository.AiChatRepository
-import com.flashcardsopensourceapp.data.local.repository.AutoSyncEventRepository
+import com.flashcardsopensourceapp.data.local.repository.sync.AutoSyncEventRepository
 import com.flashcardsopensourceapp.data.local.repository.CardsRepository
 import com.flashcardsopensourceapp.data.local.repository.cloudsync.account.CloudIdentityResetCoordinator
 import com.flashcardsopensourceapp.data.local.repository.cloudsync.guest.CloudGuestSessionCoordinator
@@ -62,17 +62,17 @@ import com.flashcardsopensourceapp.data.local.repository.DecksRepository
 import com.flashcardsopensourceapp.data.local.repository.FeedbackRepository
 import com.flashcardsopensourceapp.data.local.repository.ai.LocalAiChatRepository
 import com.flashcardsopensourceapp.data.local.repository.cloudsync.account.LocalCloudAccountRepository
-import com.flashcardsopensourceapp.data.local.repository.LocalCardsRepository
-import com.flashcardsopensourceapp.data.local.repository.LocalDecksRepository
-import com.flashcardsopensourceapp.data.local.repository.LocalFeedbackRepository
+import com.flashcardsopensourceapp.data.local.repository.cards.LocalCardsRepository
+import com.flashcardsopensourceapp.data.local.repository.decks.LocalDecksRepository
+import com.flashcardsopensourceapp.data.local.repository.feedback.LocalFeedbackRepository
 import com.flashcardsopensourceapp.data.local.repository.progress.cache.LocalProgressCacheStore
 import com.flashcardsopensourceapp.data.local.repository.progress.LocalProgressRepository
-import com.flashcardsopensourceapp.data.local.repository.LocalReviewRepository
+import com.flashcardsopensourceapp.data.local.repository.review.LocalReviewRepository
 import com.flashcardsopensourceapp.data.local.repository.cloudsync.sync.LocalSyncRepository
-import com.flashcardsopensourceapp.data.local.repository.LocalWorkspaceRepository
+import com.flashcardsopensourceapp.data.local.repository.workspace.LocalWorkspaceRepository
 import com.flashcardsopensourceapp.data.local.repository.ProgressRepository
 import com.flashcardsopensourceapp.data.local.repository.ReviewRepository
-import com.flashcardsopensourceapp.data.local.repository.SystemTimeProvider
+import com.flashcardsopensourceapp.data.local.repository.shared.SystemTimeProvider
 import com.flashcardsopensourceapp.data.local.repository.SyncRepository
 import com.flashcardsopensourceapp.data.local.repository.WorkspaceRepository
 import kotlinx.coroutines.CancellationException
