+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/repository/shared/TimeProvider.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local.repository
+package com.flashcardsopensourceapp.data.local.repository.shared
 
 import java.time.ZoneId
