+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/repository/sync/AutoSyncEvents.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local.repository
+package com.flashcardsopensourceapp.data.local.repository.sync
 
 import kotlinx.coroutines.flow.Flow
