+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/repository/workspace/WorkspaceRepositorySupport.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local.repository
+package com.flashcardsopensourceapp.data.local.repository.workspace
 
 import com.flashcardsopensourceapp.data.local.database.entities.WorkspaceSchedulerSettingsEntity
 import com.flashcardsopensourceapp.data.local.model.cards.CardSummary
